LayoutTests/fast/css/ensure-flexbox-compatibility-with-initial-values-expected.txt

Test to verify initial values of alignment properties are backward-comaptible with flexbox implementation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Verifying initial values are supported when grid is ENABLED.
PASS CSS.supports('align-items', 'normal') is true
PASS CSS.supports('align-self', 'normal') is true
PASS CSS.supports('align-content', 'normal') is true
PASS CSS.supports('justify-content', 'normal') is true
PASS CSS.supports('align-items', 'normal') is true
PASS CSS.supports('align-self', 'normal') is true
PASS CSS.supports('align-content', 'normal') is true
PASS CSS.supports('justify-content', 'normal') is true
PASS CSS.supports('align-items', 'normal') is true
PASS CSS.supports('align-self', 'normal') is true
PASS CSS.supports('align-content', 'normal') is true
PASS CSS.supports('justify-content', 'normal') is true
PASS CSS.supports('align-items', 'normal') is true
PASS CSS.supports('align-self', 'normal') is true
PASS CSS.supports('align-content', 'normal') is true
PASS CSS.supports('justify-content', 'normal') is true

Verifying initial values are supported when grid is DISABLED.
PASS CSS.supports('align-items', 'stretch') is true
PASS CSS.supports('align-self', 'stretch') is true
PASS CSS.supports('align-content', 'stretch') is true
PASS CSS.supports('justify-content', 'flex-start') is true
PASS CSS.supports('align-items', 'stretch') is true
PASS CSS.supports('align-self', 'stretch') is true
PASS CSS.supports('align-content', 'stretch') is true
PASS CSS.supports('justify-content', 'flex-start') is true
PASS CSS.supports('align-items', 'stretch') is true
PASS CSS.supports('align-self', 'stretch') is true
PASS CSS.supports('align-content', 'stretch') is true
PASS CSS.supports('justify-content', 'flex-start') is true
PASS CSS.supports('align-items', 'stretch') is true
PASS CSS.supports('align-self', 'stretch') is true
PASS CSS.supports('align-content', 'stretch') is true
PASS CSS.supports('justify-content', 'flex-start') is true
PASS successfullyParsed is true

TEST COMPLETE
